fix(proxy): apply ad-filter URL rewrite on /api/search too
Initial implementation only rewrote episodes URLs in /api/detail.
But the web frontend (src/app/play/page.tsx:2168) and OrionTV
(stores/detailStore.ts:152) both consume episodes from /api/search
and /api/search/one responses, so the rewrite never reached the
player and the proxy was effectively dead code on the playback path.

Extract the rewrite logic to src/lib/episode-rewriter.ts and apply
it on every endpoint that returns SearchResult.episodes:
  - /api/detail
  - /api/search
  - /api/search/one
  - /api/search/ws (SSE per-source streaming events)

Same skip rules as before: no-op for private_library, non-m3u8
URLs, non-http URLs, when ENABLE_AD_FILTER=false, or when caller
passes ?adfilter=false.
