+++ src/app/api/detail/route.ts
@@ -3,6 +3,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import { getAuthInfoFromCookie, verifyApiAuth } from '@/lib/auth';
 import { getAvailableApiSites, getCacheTime } from '@/lib/config';
 import { getDetailFromApi } from '@/lib/downstream';
+import { rewriteEpisodesForAdFilter } from '@/lib/episode-rewriter';
 import {
   buildPrivateLibraryPosterUrl,
   formatPrivateLibrarySourceName,
@@ -16,52 +17,6 @@ import {
 import { SearchResult } from '@/lib/types';
 
 export const runtime = 'nodejs';
-
-function isAdFilterEnabled(): boolean {
-  const flag = process.env.ENABLE_AD_FILTER;
-  if (flag === undefined) return true; // 默认开
-  return flag === 'true' || flag === '1';
-}
-
-function adFilterDisabledByQuery(request: NextRequest): boolean {
-  const v = request.nextUrl.searchParams.get('adfilter');
-  return v === 'false' || v === '0';
-}
-
-function buildFilterProxyUrl(request: NextRequest, upstreamUrl: string): string {
-  const host = request.headers.get('host');
-  const protocol =
-    request.headers.get('x-forwarded-proto') ||
-    request.nextUrl.protocol.replace(':', '') ||
-    'http';
-  return `${protocol}://${host}/api/proxy/m3u8-filter?url=${encodeURIComponent(
-    upstreamUrl,
-  )}`;
-}
-
-function shouldRewriteEpisode(url: string): boolean {
-  if (!url) return false;
-  if (!/^https?:\/\//i.test(url)) return false; // 跳过 /api/private-library/stream 这类内部路径
-  if (!/\.m3u8(\?|$|#)/i.test(url)) return false; // 只处理 m3u8
-  return true;
-}
-
-function maybeRewriteEpisodesForAdFilter(
-  result: SearchResult,
-  request: NextRequest,
-): SearchResult {
-  if (!isAdFilterEnabled() || adFilterDisabledByQuery(request)) return result;
-  if (result.source === 'private_library') return result;
-  if (!Array.isArray(result.episodes) || result.episodes.length === 0) {
-    return result;
-  }
-
-  const rewritten = result.episodes.map((ep) =>
-    shouldRewriteEpisode(ep) ? buildFilterProxyUrl(request, ep) : ep,
-  );
-
-  return { ...result, episodes: rewritten };
-}
 
 export async function GET(request: NextRequest) {
   const authResult = verifyApiAuth(request);
@@ -211,7 +166,7 @@ export async function GET(request: NextRequest) {
     const result = await getDetailFromApi(apiSite, id);
     const cacheTime = await getCacheTime();
 
-    const finalResult = maybeRewriteEpisodesForAdFilter(result, request);
+    const finalResult = rewriteEpisodesForAdFilter(result, request);
 
     return NextResponse.json(finalResult, {
       headers: {
